requirements.txt
# Core dependencies (unpinned for compatibility)
boto3
snowflake-connector-python
snowflake-connector-python[pandas]
pandas
numpy
matplotlib
seaborn
requests
tqdm

# Data quality analysis dependencies
regex
# langdetect  # Removed due to import issues
# lingua-py  # Commented out due to yanked versions - use alternative approach
